feat(con-126): add ref and onClick func for tracking
Renamed component to tsx to address forwardRef issue
